Cleanup, use absolute positioning to stick icon in top left
