+++ styled/checkout.js
@@ -86,34 +86,3 @@ export const RewardAppliedContainer = styled(FilledButtonContainer)`
   border-radius: 9px;
   margin: 8px 0px;
 `;
-
-// Modals
-
-export const ModalContentContainer = styled(ColumnContainer)`
-  height: 40%;
-  width: 45%;
-  margin: auto;
-  justify-content: space-around;
-  align-items: center;
-  background-color: white;
-  top: -200px;
-`;
-
-export const QuantityInput = styled.TextInput`
-  border-width: 1px;
-  border-color: ${Colors.activeText};
-  height: 40px;
-  width: 200px;
-  color: ${Colors.activeText};
-  font-size: 18px;
-  font-weight: bold;
-  text-align: center;
-`;
-
-export const ModalCenteredOpacityLayer = styled(RowContainer)`
-  height: 100%;
-  width: 100%;
-  background-color: rgba(0, 0, 0, 0.3);
-  justify-content: center;
-  align-items: center;
-`;
